pyproject.toml
[project]
name = "sistema-personal-wsp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.28",
    "openai>=1.98.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "twilio>=9.7.0",
    "uvicorn>=0.35.0",
]
